# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5fc13d99a3e04911232345d8
  title: famkit
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  "/contact/{id}":
    delete:
      x-skaffolder-id: 5fc148fba3e0491123234668
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5fc140cba3e049112323465e
      x-skaffolder-resource: Contact
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Contact
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5fc1482da3e0491123234664
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5fc140cba3e049112323465e
      x-skaffolder-resource: Contact
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Contact
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Contact
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contact:
    get:
      x-skaffolder-id: 5fc14871a3e0491123234665
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5fc140cba3e049112323465e
      x-skaffolder-resource: Contact
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF Contact
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Contact
      parameters:
        - name: userId
          x-skaffolder-type: String
          in: path
          description: user contacts belong to
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  "/user/{id}/changePassword":
    post:
      x-skaffolder-id: 5fc13d99a3e04911232345e6
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5fc13d99a3e04911232345e2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles:
        - 5fc13d99a3e04911232345e4
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
  /user:
    post:
      x-skaffolder-id: 5fc13d99a3e04911232345e5
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5fc13d99a3e04911232345e2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5fc13d99a3e04911232345ec
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5fc13d99a3e04911232345e2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5fc13d99a3e04911232345e8
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5fc13d99a3e04911232345e2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5fc13d99a3e04911232345e7
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5fc13d99a3e04911232345e2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5fc13d99a3e04911232345e9
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5fc13d99a3e04911232345e2
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Contact:
      x-skaffolder-id: 5fc140cba3e049112323465e
      x-skaffolder-id-db: 5fc13d99a3e04911232345d9
      x-skaffolder-id-entity: 5fc140cba3e049112323465c
      x-skaffolder-url: /contact
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        email:
          type: string
          x-skaffolder-id-attr: 5fc147efa3e0491123234662
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5fc140cba3e049112323465d
          x-skaffolder-type: String
          x-skaffolder-required: true
        phone:
          type: string
          x-skaffolder-id-attr: 5fc147efa3e0491123234663
          x-skaffolder-type: String
        userId:
          type: object
          x-skaffolder-id-attr: 5fc147efa3e0491123234661
          x-skaffolder-type: Custom
      x-skaffolder-relations: 
      required:
        - _id
        - name
    User:
      x-skaffolder-id: 5fc13d99a3e04911232345e2
      x-skaffolder-id-db: 5fc13d99a3e04911232345d9
      x-skaffolder-id-entity: 5fc13d99a3e04911232345da
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        email:
          type: string
          x-skaffolder-id-attr: 5fc13d99a3e04911232345dd
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5fc13d99a3e04911232345de
          x-skaffolder-type: String
          x-skaffolder-required: true
        password:
          type: string
          x-skaffolder-id-attr: 5fc13d99a3e04911232345dc
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5fc13d99a3e04911232345e0
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5fc13d99a3e04911232345df
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5fc13d99a3e04911232345db
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        contacts:
          x-skaffolder-id: 5fc145fea3e0491123234660
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5fc13d99a3e04911232345da
          x-skaffolder-ent2: 5fc140cba3e049112323465c
      required:
        - _id
        - name
        - password
        - username
  x-skaffolder-db:
    - x-skaffolder-id: 5fc13d99a3e04911232345d9
      x-skaffolder-name: famkit_db
  x-skaffolder-page:
    - x-skaffolder-id: 5fc14941a3e049112323466b
      x-skaffolder-name: ContactList
      x-skaffolder-url: /contacts
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5fc140cba3e049112323465e
      x-skaffolder-services:
        - 5fc14871a3e0491123234665
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5fc13d99a3e04911232345e1
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
  x-skaffolder-roles: 
    - x-skaffolder-id: 5fc13d99a3e04911232345e4
      x-skaffolder-name: ADMIN